% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbifDownload.R
\name{autoplot.gbifOcc}
\alias{autoplot.gbifOcc}
\title{Plot GBIF Occurrence Data}
\usage{
autoplot.gbifOcc(object, ...)
}
\arguments{
\item{object}{An object that inherits from class \code{gbifOcc} (such as that
produced by the \code{\link{gbifOccDownload}} function).}

\item{...}{Named arguments passed to constituent plotting functions.
Arguments that are of the form \code{points.parameter} will be passed to the
\code{\link[ggplot2]{geom_sf}} function that determines the plotting
behaviour of the occurrence records (without the \code{points.} prefix). If
arguments of the form \code{boundary.parameter} are provided then a boundary
layer is added to the plot (with data drawn from the \code{boundary.data}
argument) and these arguments are similarly passed to the corresponding
\code{\link[ggplot2]{geom_sf}} function (without the \code{boundary.}
prefix). Optionally an argument titled \code{titleLabelWidth} can be provided
which is an integer scalar giving the number of characters to aim for when
wrapping text associated with the titles of dataset names in legends and axis
labels.}
}
\value{
A \code{ggplot} layer.
}
\description{
Function to visualise the data associated with a GBIF occurrence
download.
}
\seealso{
\code{\link{gbifOccDownload}}, \code{\link[ggplot2]{autoplot}},
\code{\link[ggplot2]{geom_sf}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
