% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariateProcessing.R
\name{getCovariateLayer}
\alias{getCovariateLayer}
\title{Retrieve a Covariate Layer at a Given Time Stamp}
\usage{
getCovariateLayer(
  x,
  subds = 0,
  lyrs = NULL,
  drivers = NULL,
  opts = NULL,
  win = NULL,
  snap = "near",
  vsi = FALSE,
  raw = FALSE,
  noflip = FALSE,
  guessCRS = TRUE,
  domains = "",
  md = NULL,
  dims = NULL,
  lyrDate = NULL
)
}
\arguments{
\item{x}{A character scalar containing the filename of the raster file
containing the covariate information.}

\item{subds}{See \code{\link[terra]{rast}}.}

\item{lyrs}{See \code{\link[terra]{rast}}.}

\item{drivers}{See \code{\link[terra]{rast}}.}

\item{opts}{See \code{\link[terra]{rast}}.}

\item{win}{See \code{\link[terra]{rast}}.}

\item{snap}{See \code{\link[terra]{rast}}.}

\item{vsi}{See \code{\link[terra]{rast}}.}

\item{raw}{See \code{\link[terra]{rast}}.}

\item{noflip}{See \code{\link[terra]{rast}}.}

\item{guessCRS}{See \code{\link[terra]{rast}}.}

\item{domains}{See \code{\link[terra]{rast}}.}

\item{md}{See \code{\link[terra]{rast}}.}

\item{dims}{See \code{\link[terra]{rast}}.}

\item{lyrDate}{An object containing the date information of the covariate
layer.  Can be any type coercible by \code{\link[base]{as.Date}}.}
}
\value{
A \code{SpatRaster}.  If \code{lyrdate} is not \code{NULL} then the
output has an extra attribute, \code{"lyrdate"}, the contains the timestemp
information.
}
\description{
Function to load a covariate layer and (optionally) add a
timestamp to it.
}
\seealso{
\code{\link[base]{as.Date}}, \code{\link[terra]{rast}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
