% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSDM.R
\name{runSDM}
\alias{runSDM}
\title{Run a distribution model}
\usage{
runSDM(
  sdmFunc,
  occData,
  group = NULL,
  outLoc = tempdir(),
  numCores = NULL,
  ...
)
}
\arguments{
\item{sdmFunc}{A function that runs the species distribution model.}

\item{occData}{An \code{\link[sf]{sf}} object containing the occurrence data
of the taxa to be modelled.}

\item{group}{A character scalar that is used identify the column name in
\code{occData} that is used to assign the different batches that the species
distribution model is going to be applied.  If \code{NULL} (the default),
then only a single SDM is run on the entire occurrence data. If a valid
column name is given then a separate SDM is run for each subset of the
occurrence data based on unique values in that column.}

\item{outLoc}{A character scalar containing the location to store the output
model objects. If \code{group} is \code{NULL} then the model outputs are
stored directly in the directory specified. Otherwise, a sub-folder is
created for each SDM.}

\item{numCores}{Integer vector containing the number of processes to be run
simultaneously. If \code{group} is not \code{NULL} then the first element of
the integer vector is used to divide resources for processing each of the
SDMs for the groups.  All unused elements of the integer vector are passed to
\code{sdmFunc} (and may be used there to control parallelisation within the
SDM function). If \code{NULL}, then defaults to the number of cores
available.}

\item{...}{Named arguments to be passed to \code{sdmFunc}.}
}
\value{
A logical vector giving the exit status of the model functions.
}
\description{
This function runs a single species distribution model (or series of species
distribution models if the \code{group} parameter is not \code{NULL}).
The function is also responsible for collating the relevant metadata for the
analysis.
}
\seealso{
\code{\link[terra]{sds}}, \code{\link[terra]{rast}},
\code{\link[base]{strptime}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
