% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbifDownload.R
\name{taxonLookup}
\alias{taxonLookup}
\title{Lookup GBIF Taxonomic Information}
\usage{
taxonLookup(
  kingdom = NULL,
  phylum = NULL,
  class = NULL,
  order = NULL,
  family = NULL,
  genus = NULL,
  species = NULL,
  other = NULL,
  limit = NULL,
  rank = NULL,
  synonyms = NULL,
  ...
)
}
\arguments{
\item{kingdom}{A character vector containing the kingdoms that the taxa
retrieved from GBIF's taxonomies database must belong to. An entry with a
length greater than one will ensure that all retrieved entries must belong
to one of the kingdoms specified in the input. A value of \code{NULL} ensures
that entries will not be filtered according to any criteria at this taxonomic
level.}

\item{phylum}{A character vector containing the phyla that the taxa
retrieved from GBIF's taxonomies database must belong to. An entry with a
length greater than one will ensure that all retrieved entries must belong
to one of the phyla specified in the input. A value of \code{NULL} ensures
that entries will not be filtered according to any criteria at this taxonomic
level.}

\item{class}{A character vector containing the classes that the taxa
retrieved from GBIF's taxonomies database must belong to. An entry with a
length greater than one will ensure that all retrieved entries must belong
to one of the classes specified in the input. A value of \code{NULL} ensures
that entries will not be filtered according to any criteria at this taxonomic
level.}

\item{order}{A character vector containing the orders that the taxa
retrieved from GBIF's taxonomies database must belong to. An entry with a
length greater than one will ensure that all retrieved entries must belong
to one of the orders specified in the input. A value of \code{NULL} ensures
that entries will not be filtered according to any criteria at this taxonomic
level.}

\item{family}{A character vector containing the families that the taxa
retrieved from GBIF's taxonomies database must belong to. An entry with a
length greater than one will ensure that all retrieved entries must belong
to one of the families specified in the input. A value of \code{NULL} ensures
that entries will not be filtered according to any criteria at this taxonomic
level.}

\item{genus}{A character vector containing the genera that the taxa
retrieved from GBIF's taxonomies database must belong to. An entry with a
length greater than one will ensure that all retrieved entries must belong
to one of the genera specified in the input. A value of \code{NULL} ensures
that entries will not be filtered according to any criteria at this taxonomic
level.}

\item{species}{A character vector containing the species that the taxa
retrieved from GBIF's taxonomies database must belong to. An entry with a
length greater than one will ensure that all retrieved entries must belong
to one of the species specified in the input. A value of \code{NULL} ensures
that entries will not be filtered according to any criteria at this taxonomic
level.}

\item{other}{A character vector containing the values for the taxonomic level
specified in the \code{rank} argument that the taxa retrieved from GBIF's
taxonomies database must belong to. An entry with a length greater than one
will ensure that all retrieved entries must belong to one of the values
specified in the input. A value of \code{NULL} ensures that entries will not
be filtered according to any criteria at this taxonomic level.}

\item{limit}{An integer scalar containing the maximum number of entries to
return from the query. A value of \code{NULL} indicates that all valid
entries will be returned.}

\item{rank}{If the \code{other} parameter is non-\code{NULL} then a scalar
character must be provided here defining the taxonomic level to conduct the
query at. See \code{\link[rgbif]{name_lookup}} for acceptable values to use
for the \code{rank} argument.}

\item{synonyms}{A character vector containing the list of taxonomic databases
to search for possible synonym names for the taxa names. See "details" for
more information on this.}

\item{...}{Other parameters passed to the \code{\link[rgbif]{name_lookup}}
function when querying GBIF's taxonomies database.}
}
\value{
A \code{data.frame} containing the taxonomic information for all
matching taxa according to the criteria specified in the function arguments.
}
\description{
A function that iteratively calls the
\code{\link[rgbif]{name_lookup}} function to retrieve information relating to
a taxon (or selection of taxa) contained at \href{https://www.gbif.org/}{GBIF}.
}
\details{
The \code{\link[rgbif]{name_lookup}} functions that is used in this
function has been flagged as deprecated. However, the suggested replacements
for this function are in a package that has been removed from CRAN and
requires installation through a separate package manager.  Using the
replacement would therefore induce complex dependencies in the ECoMAPr
package if we were to use that instead. We will therefore keep the dependence
upon the deprecated function for the time being but we will update the
internal implementation of this function when CRAN-compliant options become
available.

The \code{synonyms} argument is a character vector containing the set of
taxonomic databases to check for possible synonyms. Each element of the
character may have the following (case-insensitive) values: \code{"gbif"}
(Global Biodiversity Information Facility taxonomic backbone), \code{"itis"}
(Integrated Taxonomic Information System), \code{"tropicos"} (Tropicos online
database of the Missouri Botanical Garden), \code{"nbn"} ("National
Biodiversity Network"), \code{"worms"} ("World Register of Marine Species"),
\code{"pow"} ("Plants of the World Online"). If the argument is NULL then the
function will default to using all available databases.  Note that use of the
Tropicos database requires an API key: see \code{\link[taxize]{synonyms}} for
more information.
}
\seealso{
\code{\link[rgbif]{name_lookup}} \code{\link[rgbif]{name_usage}}
\code{\link[taxize]{synonyms}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
