% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occProcessing.R
\name{classifyOccDataType}
\alias{classifyOccDataType}
\title{Classify Occurrence Data}
\usage{
classifyOccDataType(
  occData,
  datasetIDCol = "datasetKey",
  occurrenceStatusCol = "occurrenceStatus",
  densityCol = c("individualCount", "organismQuantity"),
  occurrenceOutColName = "occValues",
  densityOutColName = "densityValues"
)
}
\arguments{
\item{occData}{A data frame (or object that inherits from a data frame)
containing the occurrence data to classify.}

\item{datasetIDCol}{A character vector containing the name of the column
containing a the dataset ID grouping variable.}

\item{occurrenceStatusCol}{A character vector containing the name of the
column(s) containing the occurrence status (marked as \code{"PRESENT"} or
\code{"ABSENT"}).}

\item{densityCol}{A character vector containing the name of the column(s)
containing the density data.}

\item{occurrenceOutColName}{A character scalar giving the name for the extra
column that will be appended to the occurrence data that will store the
combined occurrence information.}

\item{densityOutColName}{A character scalar giving the name for the extra
column that will be appended to the density data that will store the combined
density information.}
}
\value{
A list with the following elements:
\describe{
\item{\code{presenceOnly}}{A data frame containing the rows that have been
classified as having data that only contains presence records with no
information on sampling effort ("presence only data").}
\item{\code{presenceAbsence}}{A data frame containing the rows that have
been classified as having data that contains presences or absences
("presence/absence data"). An extra column is appended to the data frame
with a name specified by \code{occurrenceOutColName}.}
\item{\code{density}}{A data frame containing the rows that have been
classified as having data that contains some form of density information
("density data").  An extra column is appended to the data frame with a name
specified by \code{densityOutColName}.}
}
}
\description{
Function to take species occurrence data and classify it into
categories of 'presence only', 'presence/absence', or 'density'.
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
